container/requirements.txt
langchain
langchain-community 
langchain-core
chromadb

langchain-openai
fastembed
faiss-cpu  # Para manejar la indexación en el RAG
